refactor(core): extract shared component scope
