+++ README.md
@@ -64,7 +64,7 @@ The current build enables the first lazy modules:
 - `kotventure-test` — Kotest component matchers consumed test-scoped by library modules
 - `kotventure-bom` — a Gradle/Maven BOM aligning enabled Kotventure artifacts and Adventure 5.1.1 dependencies
 
-The `core` module exposes a composable text builder:
+The `core` module exposes a composable component tree builder:
 
 ```kotlin
 val badge = Component.text("[new]", NamedTextColor.GREEN)
@@ -96,11 +96,12 @@ val itemCount = translatable("item.count") {
 }
 ```
 
-Smaller Adventure component kinds are available directly too:
+Smaller Adventure component kinds are available directly and inside any component scope:
 
 ```kotlin
 val jumpHint = keybind("key.jump") {
     color(NamedTextColor.YELLOW)
+    text(" to jump")
 }
 
 val killCount = score("Alex", "kills")
