+++ modules/core/src/main/kotlin/io/github/lmliam/kotventure/core/score/ScoreComponentBuilder.kt
@@ -1,57 +1,13 @@
 package io.github.lmliam.kotventure.core.score
 
-import io.github.lmliam.kotventure.core.style.StyleScope
+import io.github.lmliam.kotventure.core.component.AbstractComponentScopeBuilder
 import net.kyori.adventure.text.Component
 import net.kyori.adventure.text.ScoreComponent
-import net.kyori.adventure.text.format.Style
-import net.kyori.adventure.text.format.TextColor
-import net.kyori.adventure.text.format.TextDecoration
 
 internal class ScoreComponentBuilder(
     name: String,
     objective: String,
-) : ScoreScope {
-    private val builder: ScoreComponent.Builder = Component.score().name(name).objective(objective)
-
-    override fun color(color: TextColor) {
-        builder.color(color)
-    }
-
-    override fun style(style: Style) {
-        builder.style(style)
-    }
-
-    override fun style(init: StyleScope.() -> Unit) {
-        ScoreStyleScope(builder).init()
-    }
-
-    override fun decorate(decoration: TextDecoration) {
-        builder.decoration(decoration, true)
-    }
-
-    override fun bold() {
-        decorate(TextDecoration.BOLD)
-    }
-
-    override fun italic() {
-        decorate(TextDecoration.ITALIC)
-    }
-
-    override fun underlined() {
-        decorate(TextDecoration.UNDERLINED)
-    }
-
-    override fun strikethrough() {
-        decorate(TextDecoration.STRIKETHROUGH)
-    }
-
-    override fun obfuscated() {
-        decorate(TextDecoration.OBFUSCATED)
-    }
-
-    override fun append(component: Component) {
-        builder.append(component)
-    }
-
-    internal fun build(): Component = builder.build()
-}
+) : AbstractComponentScopeBuilder<ScoreComponent, ScoreComponent.Builder>(
+    Component.score().name(name).objective(objective),
+),
+    ScoreScope
