+++ modules/core/src/main/kotlin/io/github/lmliam/kotventure/core/text/Component.kt
@@ -1,8 +1,9 @@
 package io.github.lmliam.kotventure.core.text
 
+import io.github.lmliam.kotventure.core.component.ComponentScope
 import net.kyori.adventure.text.Component
 
 /**
- * Builds an Adventure [Component] directly from a Kotventure text DSL block.
+ * Builds an Adventure [Component] directly from a Kotventure component DSL block.
  */
-public fun component(init: TextScope.() -> Unit): Component = TextComponentBuilder().apply(init).build()
+public fun component(init: ComponentScope.() -> Unit): Component = TextComponentBuilder().apply(init).build()
